Check 'git status vendor/' in verify-vendor.sh
If vendor/ has new uncommitted files, verify-vendor.sh will not detect
it, and this can be quite puzzling when this check fails on CI.
Checking 'git status vendor/' will make this situation easier to detect.

+++ verify-vendor.sh
@@ -17,6 +17,11 @@ trap "cleanup" EXIT SIGINT
 
 cleanup
 
+if [[ -n $(git status -s vendor/) ]]; then
+        echo 'vendor/ directory has uncommitted changes, please check `git status vendor`'
+        exit 1
+fi
+
 mkdir "${TMP_DIFFROOT}"
 cp -aR "${REPO_ROOT_DIR}/vendor" "${TMP_DIFFROOT}"
